pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "urpan-inpaint"
version = "0.1.0"
description = "Production pipeline scaffold for 360 street-scene inpainting and alpha compositing"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "urpan-inpaint"}]
dependencies = [
  "numpy>=1.24",
  "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0"
]

[project.scripts]
urpan-inpaint = "urpan_inpaint.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
